autolav-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
playwright==1.48.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-dateutil==2.9.0
openpyxl==3.1.5
pytest==8.3.3
pytest-asyncio==0.24.0
aiofiles==24.1.0
